Move and clean up all utils code
