+++ Assets/Scripts/Player/InventorySlotHUD.cs
@@ -2,6 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
+using DeCrawl.Utils;
 
 public enum InventorySlotHUDState { Corrupted, Free, Occupied };
