+++ Assets/Scripts/Systems/PositionRecorder.cs
@@ -2,7 +2,7 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using System.Linq;
-using CompressString;
+using DeCrawl.Utils;
 
 public class PositionRecorder : MonoBehaviour, StateSaver
 {
