+++ Assets/deCrawl.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 70cdf34cfea16eb4c8009a9e028f73d3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
