+++ Assets/deCrawl/Scrips.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8af3ed66eafae3243851db431a688f36
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
