+++ Assets/deCrawl/Scrips/Utils/LineSearch.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 075f49d32c0e52540960788729d20af5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
